+++ modules/Sources/Dependencies/SDKSynchronizer/SDKSynchronizerTest.swift
@@ -30,7 +30,10 @@ extension SDKSynchronizerClient: TestDependencyKey {
         sendTransaction: XCTUnimplemented("\(Self.self).sendTransaction", placeholder: .placeholder()),
         shieldFunds: XCTUnimplemented("\(Self.self).shieldFunds", placeholder: .placeholder()),
         wipe: XCTUnimplemented("\(Self.self).wipe"),
-        switchToEndpoint: XCTUnimplemented("\(Self.self).switchToEndpoint")
+        switchToEndpoint: XCTUnimplemented("\(Self.self).switchToEndpoint"),
+        proposeTransfer: XCTUnimplemented("\(Self.self).proposeTransfer", placeholder: .testOnlyFakeProposal(totalFee: 0)),
+        createProposedTransactions: XCTUnimplemented("\(Self.self).createProposedTransactions", placeholder: .success),
+        proposeShielding: XCTUnimplemented("\(Self.self).proposeShielding", placeholder: nil)
     )
 }
 
@@ -52,7 +55,10 @@ extension SDKSynchronizerClient {
         sendTransaction: { _, _, _, _ in return .placeholder() },
         shieldFunds: { _, _, _ in return .placeholder() },
         wipe: { Empty<Void, Error>().eraseToAnyPublisher() },
-        switchToEndpoint: { _ in }
+        switchToEndpoint: { _ in },
+        proposeTransfer: { _, _, _, _ in .testOnlyFakeProposal(totalFee: 0) },
+        createProposedTransactions: { _, _ in .success },
+        proposeShielding: { _, _, _, _ in nil }
     )
 
     public static let mock = Self.mocked()
@@ -172,7 +178,13 @@ extension SDKSynchronizerClient {
             )
         },
         wipe: @escaping () -> AnyPublisher<Void, Error>? = { Fail(error: "Error").eraseToAnyPublisher() },
-        switchToEndpoint: @escaping (LightWalletEndpoint) async throws -> Void = { _ in }
+        switchToEndpoint: @escaping (LightWalletEndpoint) async throws -> Void = { _ in },
+        proposeTransfer: 
+        @escaping (Int, Recipient, Zatoshi, Memo?) async throws -> Proposal = { _, _, _, _ in .testOnlyFakeProposal(totalFee: 0) },
+        createProposedTransactions:
+        @escaping (Proposal, UnifiedSpendingKey) async throws -> CreateProposedTransactionsResult = { _, _ in .success },
+        proposeShielding:
+        @escaping (Int, Zatoshi, Memo, TransparentAddress?) async throws -> Proposal? = { _, _, _, _ in nil }
     ) -> SDKSynchronizerClient {
         SDKSynchronizerClient(
             stateStream: stateStream,
@@ -191,7 +203,10 @@ extension SDKSynchronizerClient {
             sendTransaction: sendTransaction,
             shieldFunds: shieldFunds,
             wipe: wipe,
-            switchToEndpoint: switchToEndpoint
+            switchToEndpoint: switchToEndpoint,
+            proposeTransfer: proposeTransfer,
+            createProposedTransactions: createProposedTransactions,
+            proposeShielding: proposeShielding
         )
     }
 }