+++ secantTests/SnapshotTests/BalanceBreakdownSnapshotTests/BalanceBreakdownSnapshotTests.swift
@@ -10,6 +10,7 @@ import ComposableArchitecture
 import ZcashLightClientKit
 import SwiftUI
 import BalanceBreakdown
+import PartialProposalError
 @testable import secant_testnet
 
 class BalanceBreakdownSnapshotTests: XCTestCase {
@@ -19,6 +20,7 @@ class BalanceBreakdownSnapshotTests: XCTestCase {
                 autoShieldingThreshold: Zatoshi(1_000_000),
                 changePending: .zero,
                 isShieldingFunds: false,
+                partialProposalErrorState: .initial,
                 pendingTransactions: .zero,
                 shieldedBalance: Zatoshi(123_000_000_000),
                 syncProgressState: .initial,
@@ -42,6 +44,7 @@ class BalanceBreakdownSnapshotTests: XCTestCase {
                 changePending: .zero,
                 isShieldingFunds: false,
                 isHintBoxVisible: true,
+                partialProposalErrorState: .initial,
                 pendingTransactions: .zero,
                 shieldedBalance: Zatoshi(123_000_000_000),
                 syncProgressState: .initial,